+++ tests/torch2/fx/test_compress_weights.py
@@ -19,7 +19,10 @@ from nncf import CompressWeightsMode
 from nncf import SensitivityMetric
 from nncf.common.factory import NNCFGraphFactory
 from nncf.experimental.torch.fx.node_utils import get_tensor_constant_from_node
+from nncf.experimental.torch.fx.node_utils import get_reduction_axes_from_metatype
+from nncf.torch.graph import operator_metatypes as om
 from nncf.experimental.torch.fx.transformations import get_graph_node_by_name
+from nncf.quantization.algorithms.weight_compression.torch_fx_backend import FXWeightCompressionAlgoBackend
 from nncf.parameters import CompressionFormat
 from nncf.quantization import compress_weights
 from nncf.quantization.advanced_parameters import AdvancedCompressionParameters
@@ -313,6 +316,39 @@ def test_model_devices_and_precisions(use_cuda, dtype):
     # Result should be in the precision of the model
     assert result.dtype == dtype
 
+@pytest.mark.parametrize(
+    "metatype,weight_port_id,ndims,expected_reduction_axes",
+    [
+        (om.PTAtenEmbeddingMetatype, 0, 4, [1]),
+        (om.PTLinearMetatype,        0, 4, [3]),
+        (om.PTMatMulMetatype,        0, 4, [3]),
+        (om.PTMatMulMetatype,        1, 4, [2]),
+        (om.PTAddmmMetatype,         1, 4, [3]),
+        (om.PTAddmmMetatype,         2, 4, [2]),
+        (om.PTMatMulMetatype,        1, 1, [0]),
+        (om.PTMatMulMetatype,        1, 2, [0]),
+    ],
+)
+def test_basic_mappings(metatype, weight_port_id, ndims, expected_reduction_axes):
+    reduction_axes = get_reduction_axes_from_metatype(metatype, weight_port_id, ndims)
+    assert reduction_axes == expected_reduction_axes
+
+def _is_transposed_conv(metatype):
+    return metatype in [
+        om.PTConvTranspose1dMetatype,
+        om.PTConvTranspose2dMetatype,
+        om.PTConvTranspose3dMetatype,
+    ]
+
+def test_convolution_metatypes_reduce_all_but_channel(ndims_default):
+    ndims = ndims_default
+    conv_metatypes = list(FXWeightCompressionAlgoBackend.CONVOLUTION_METATYPES)
+
+    for metatype in conv_metatypes:
+        channel_idx = 1 if _is_transposed_conv(metatype) else 0
+        expected = [i for i in range(ndims) if i != channel_idx]
+        out = get_reduction_axes_from_metatype(mt, 0, ndims)
+        assert out == expected, f"Failed for {metatype} with ndims={ndims}"
 
 class TestFXTemplateWeightCompression(TemplateWeightCompression):
     @staticmethod
